+++ js/app/mediabunny.js
@@ -4,6 +4,7 @@ import {
     BlobSource,
     ALL_FORMATS,
     QUALITY_VERY_HIGH,
+    QUALITY_HIGH,
     Output,
     Mp4OutputFormat,
     BufferTarget,
@@ -15,9 +16,47 @@ export { getEncodableVideoCodecs } from '../../node_modules/mediabunny/dist/bund
 
 /**
  *
+ * @param {string} videoQuality
+ * @returns
+ */
+const getBitrate = videoQuality => {
+    switch (videoQuality) {
+        case 'very_low':
+            return QUALITY_HIGH;
+        default:
+            return QUALITY_VERY_HIGH;
+    }
+};
+
+/**
+ *
+ * @param {string} videoQuality
+ * @param {number|undefined} outputFps
+ * @returns {number}
+ */
+const getKeyFrameInterval = (videoQuality, outputFps = 240) => {
+    switch (videoQuality) {
+        case 'ultra':
+            return 1 / outputFps;
+        case 'very_high':
+            return Math.min((1 / outputFps) * 2, 1);
+        case 'high':
+            return Math.min((1 / outputFps) * 4, 1);
+        case 'medium':
+            return Math.min((1 / outputFps) * 8, 1);
+        case 'low':
+            return 1;
+        default:
+            return 2;
+    }
+};
+
+/**
+ *
  * @param {File} videoFile
  * @param {string} codec
  * @param {number} outputFps
+ * @param {string} videoQuality
  * @param {Function} processCallback
  * @returns
  */
@@ -25,6 +64,7 @@ export const initConversion = (
     videoFile,
     codec,
     outputFps,
+    videoQuality,
     processCallback
 ) => {
     const input = new Input({
@@ -47,7 +87,8 @@ export const initConversion = (
                 codec,
                 frameRate: outputFps,
                 forceTranscode: true,
-                bitrate: QUALITY_VERY_HIGH,
+                bitrate: getBitrate(videoQuality),
+                keyFrameInterval: getKeyFrameInterval(videoQuality, outputFps),
             },
         }),
     ];
@@ -71,12 +112,14 @@ export const createOutput = finalizeCallback => {
  * @param {HTMLCanvasElement} canvas
  * @param {string} codec
  * @param {number} outputFps
+ * @param {string} videoQuality
  * @returns
  */
-export const createCanvasSource = (canvas, codec, outputFps) => {
+export const createCanvasSource = (canvas, codec, outputFps, videoQuality) => {
     return new CanvasSource(canvas, {
         codec,
-        bitrate: QUALITY_VERY_HIGH,
+        bitrate: getBitrate(videoQuality),
+        keyFrameInterval: getKeyFrameInterval(videoQuality, outputFps),
         sizeChangeBehavior: 'contain',
         transform: {
             frameRate: outputFps,